+++ src/dropdown/bs-dropdown-toggle.directive.ts
@@ -3,10 +3,9 @@ import {
   ElementRef,
   HostBinding,
   HostListener,
-  Input,
   OnDestroy
 } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 import { BsDropdownState } from './bs-dropdown.state';