+++ src/dropdown/bs-dropdown.directive.ts
@@ -1,10 +1,18 @@
 // tslint:disable:max-file-line-count
 import {
-  Directive, ElementRef, EmbeddedViewRef, EventEmitter, Input, OnDestroy,
-  OnInit, Output, Renderer2, ViewContainerRef
+  Directive,
+  ElementRef,
+  EmbeddedViewRef,
+  EventEmitter,
+  Input,
+  OnDestroy,
+  OnInit,
+  Output,
+  Renderer2,
+  ViewContainerRef
 } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
-import 'rxjs/add/operator/filter';
+import { filter } from 'rxjs/operators';
+import { Subscription } from 'rxjs';
 import { ComponentLoader, ComponentLoaderFactory } from '../component-loader/index';
 
 import { BsDropdownConfig } from './bs-dropdown.config';
@@ -176,7 +184,9 @@ export class BsDropdownDirective implements OnInit, OnDestroy {
     // hide dropdown if set disabled while opened
     this._subscriptions.push(
       this._state.isDisabledChange
-        .filter((value: boolean) => value)
+        .pipe(
+          filter((value: boolean) => value)
+        )
         .subscribe((value: boolean) => this.hide())
     );
   }